fixup: feedback - move AUTH_DISABLED to constants
