+++ web/src/constants.ts
@@ -1,0 +1,1 @@
+export const AUTH_DISABLED = process.env.REACT_APP_AUTH !== '1'
